src/backend/requirements.txt
alembic==1.8.1
anyio==3.6.1
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coverage==6.4.2
fastapi==0.79.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.2.0
sqladmin==0.1.12
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
WTForms==3.0.1
